DIPLOM/requirements.txt
selenium==4.27.1
allure-pytest==2.13.5
pytest==8.3.3
